debugged hiseq change names
still needs to log the changes

what happened

[ticket: X]

+++ scripts/change_names_hiseq.py
@@ -1,19 +1,19 @@
 #import sys
 import os
-#import gzip
+import gzip
 import argparse
 import shutil
-
+import glob
 parser = argparse.ArgumentParser(description='This program takes hiseq fastq files and renames them as sample.read_direction.#.fastq and keeps a log of the change, it may need to be adjusted if the sequencing core changes their scheme')
 parser.add_argument('-s',action='store',dest='s',help='The sorce directory containing the original fastq files')
-parser.add_argument('-f',action='store',dest='o',help='The final directory that will hold the renamed fastq files')
+parser.add_argument('-f',action='store',dest='f',help='The final directory that will hold the renamed fastq files')
 parser.add_argument('-k',action='store',dest='key',help='The output csv given by the sequencing core that serves as the key for renaming')
 parser.add_argument('-run',action='store_true',dest='test',default=False,help='Boolean switch to run program, without this the program runs in test mode: the log is made but no files are renamed')
 
 args=parser.parse_args()
 s=args.s
 f=args.f
-key=arg.key
+key=args.key
 test=args.test
 
 if not os.path.exists(f):
@@ -34,15 +34,17 @@ if s[-1] != '/':
 junk_names = [] # The names given by the sequencing core sampleid_index
 new_names = []
 # add the bad names from the sample sheet to a list
-f =  open(key,"r")
-next(f)
-for line in f:
+names =  open(key,"r")
+next(names)
+for line in names:
     line = line.strip()
     line = line.split(',')
-    junk_names.append(line[1]+"_"+line[3])
-    new=line[4]#.replace("_","-")
+    #junk_names.append(line[6]+"_"+line[7])
+    junk_names.append(line[2])    
+    new=line[9]#.replace("_","-")
     new_names.append(new)
-f.close()
+    #print(line[9])
+names.close()
 
 if test==False:
     print "running in test mode add option -run to run"
@@ -50,7 +52,7 @@ if test==False:
 outfile = open(f+'renaming_log.txt','w')
 
 for filename in glob.glob(s + "*.fastq"):
-    name=filename.split("_L")
+    name=filename.split("_S")
     bad_name = name[0]
     lane_junk = name[1]
     read_number=lane_junk.split("_R")
@@ -61,15 +63,17 @@ for filename in glob.glob(s + "*.fastq"):
         better_name= new_names[name_index]
         perfect_name= better_name+"."+read_number+"."+fastq_number+".fastq"
         # Write file to new name
-        print("COPYING "+ s+filename + " to "+o+perfect_name)
-        outfile.write(s+filename + "\t COPIED to \t" + o+perfect_name + "\n")
+        print("COPYING "+ s+filename + " to "+f+perfect_name)
+        outfile.write(s+filename + "\t COPIED to \t" + f+perfect_name + "\n")
         if test==True:
             shutil.copy(s+filename,f+perfect_name)
             
+print(junk_names)
 for zipfilename in glob.glob(s + "*.fastq.gz"):
     zipfilename= os.path.basename(zipfilename)
-    name=filename.split("_L")
+    name=zipfilename.split("_S")
     bad_name = name[0]
+    print(bad_name)
     lane_junk = name[1]
     read_number=lane_junk.split("_R")
     fastq_number=read_number[1][4]
@@ -77,24 +81,25 @@ for zipfilename in glob.glob(s + "*.fastq.gz"):
     if bad_name in junk_names:
         name_index = junk_names.index(bad_name)
         better_name= new_names[name_index]
-        perfect_name= better_name+"."+read_number+"."+fastq_number+".fastq"
+        perfect_name= better_name+"."+read_number+"."+fastq_number+".fastq.gz"
         # Write file to new name
-        print("COPYING "+ s+filename + " to "+o+perfect_name)
-        outfile.write(s+filename + "\t COPIED to \t" + o+perfect_name + "\n")
+        print("COPYING "+ s+zipfilename + " to "+f+perfect_name +".gz")
+        outfile.write(s+zipfilename + "\t COPIED to \t" + f+perfect_name + ".gz\n")
         if test==True:
-            shutil.copy(s+filename,f+perfect_name)
+            shutil.copy(s+zipfilename,f+perfect_name)
             
-for zipfile in glob.glob(f + "*.gz"):
-    print "unzipping:" + zipfile
+if test==True:
+    for zipfile in glob.glob(f + "*.gz"):
+        print "unzipping:" + zipfile
 
-    inF = gzip.GzipFile(zipfile, 'rb')
-    s = inF.read()
-    inF.close()
-    #print(os.path.splitext(zipfile)[0])
-    outF = file(os.path.splitext(zipfile)[0], 'wb')
-    outF.write(s)
-    outF.close()
-    os.remove(zipfile)
+        inF = gzip.GzipFile(zipfile, 'rb')
+        s = inF.read()
+        inF.close()
+        #print(os.path.splitext(zipfile)[0])
+        outF = file(os.path.splitext(zipfile)[0], 'wb')
+        outF.write(s)
+        outF.close()
+        os.remove(zipfile)
 
 
 outfile.close()
